# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Take Home Test API",
      "description": "Documentation for Take Home Test API, Bambang Rahmat (NodeJs)",
      "version": "1.0.0"
    },
...
    "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "securityDefinitions": {
      "bearerAuth": {
          "type": "apiKey",
          "in": "header",
          "name": "Authorization"
      }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
    "paths": {
...
      "/balance": {
        "get": {
          "tags": ["3. Transaction"],
          "summary": "Get user balance",
          "description": "API Balance Private (memerlukan Token untuk mengaksesnya)\n\nDigunakan untuk mendapatkan informasi balance / saldo terakhir dari User.\n\nKetentuan:\n\n- Service ini harus menggunakan Bearer Token JWT untuk mengaksesnya.\n- Tidak ada parameter email di query param URL ataupun request body, parameter email diambil dari payload JWT yang didapatkan dari hasil login.\n- Handling Response sesuai dokumentasi Response di bawah.",
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Get Balance / Saldo Berhasil",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 0
                      },
                      "message": {
                        "type": "string",
                        "example": "Get Balance Berhasil"
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "balance": {
                            "type": "integer",
                            "example": 1000000
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 108
                      },
                      "message": {
                        "type": "string",
                        "example": "Token tidak valid atau kadaluwarsa"
                  },
                  "data": {
                    "type": "null"
                  }
                }
              }
            }
          }
        }
      }
        }
      },
      "/topup": {
        "post": {
          "tags": ["3. Transaction"],
          "summary": "Top up balance",
          "description": "API Topup Private (memerlukan Token untuk mengaksesnya)\n\nDigunakan untuk melakukan top up balance / saldo dari User.\n\nKetentuan:\n\n- Service ini harus menggunakan Bearer Token JWT untuk mengaksesnya.\n- Tidak ada parameter email di query param URL ataupun request body, parameter email diambil dari payload JWT yang didapatkan dari hasil login.\n- Setiap kali melakukan Top Up maka balance / saldo dari User otomatis bertambah.\n- Parameter amount hanya boleh angka saja dan tidak boleh lebih kecil dari 0.\n- Pada saat Top Up set transaction_type di database menjadi TOPUP.\n- Handling Response sesuai dokumentasi Response di bawah.",
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "top_up_amount": {
                      "type": "integer",
                      "example": 1000000
                    }
                  },
                  "required": [
                    "top_up_amount"
                  ]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Request Successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 0
                      },
                      "message": {
                        "type": "string",
                        "example": "Top Up Balance berhasil"
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "balance": {
                            "type": "integer",
                            "example": 2000000
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 102
                      },
                      "message": {
                        "type": "string",
                        "example": "Parameter amount hanya boleh angka dan tidak boleh lebih kecil dari 0"
                      },
                      "data": {
                        "type": "null"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 108
                      },
                      "message": {
                        "type": "string",
                        "example": "Token tidak valid atau kadaluwarsa"
                      },
                      "data": {
                        "type": "null"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/transaction": {
        "post": {
          "tags": ["3. Transaction"],
          "summary": "Process a transaction for available services",
          "description": "API Transaction Private (memerlukan Token untuk mengaksesnya)\n\nDigunakan untuk melakukan transaksi dari services / layanan yang tersedia.\n\nKetentuan:\n\n- Service ini harus menggunakan Bearer Token JWT untuk mengaksesnya.\n- Tidak ada parameter email di query param URL ataupun request body, parameter email diambil dari payload JWT yang didapatkan dari hasil login.\n- Setiap kali melakukan Transaksi harus dipastikan balance / saldo mencukupi.\n- Pada saat Transaction set transaction_type di database menjadi PAYMENT.\n- Handling Response sesuai dokumentasi Response di bawah.\n- Response invoice_number untuk formatnya generate bebas.",
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "service_code": {
                      "type": "string",
                      "example": "PULSA"
                    }
                  },
                  "required": [
                    "service_code"
                  ]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Transaksi Berhasil",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 0
                      },
                      "message": {
                        "type": "string",
                        "example": "Transaksi berhasil"
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "invoice_number": {
                            "type": "string",
                            "example": "INV17082023-001"
                          },
                          "service_code": {
                            "type": "string",
                            "example": "PLN_PRABAYAR"
                          },
                          "service_name": {
                            "type": "string",
                            "example": "PLN Prabayar"
                          },
                          "transaction_type": {
                            "type": "string",
                            "example": "PAYMENT"
                          },
                          "total_amount": {
                            "type": "integer",
                            "example": 10000
                          },
                          "created_on": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2023-08-17T10:10:10.000Z"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 102
                      },
                      "message": {
                        "type": "string",
                        "example": "Service atau Layanan tidak ditemukan"
                      },
                      "data": {
                        "type": "null"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 108
                      },
                      "message": {
                        "type": "string",
                        "example": "Token tidak valid atau kadaluwarsa"
                      },
                      "data": {
                        "type": "null"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/transaction/history": {
        "get": {
          "tags": ["3. Transaction"],
          "summary": "Get transaction history",
          "description": "API History Private (memerlukan Token untuk mengaksesnya)\n\nDigunakan untuk mendapatkan informasi history transaksi.\n\nKetentuan:\n\n- Service ini harus menggunakan Bearer Token JWT untuk mengaksesnya.\n- Tidak ada parameter email di query param URL ataupun request body, parameter email diambil dari payload JWT yang didapatkan dari hasil login.\n- Terdapat parameter limit yang bersifat opsional, jika limit tidak dikirim maka tampilkan semua data.\n- Data di order dari yang paling baru berdasarkan transaction date (created_on).",
          "security": [
            {
              "BearerAuth": []
            }
          ],
          "parameters": [
            {
              "name": "offset",
              "in": "query",
              "description": "Offset untuk pagination",
              "required": false,
              "schema": {
                "type": "integer",
                "example": 0
              }
            },
            {
              "name": "limit",
              "in": "query",
              "description": "Limit untuk jumlah data yang ditampilkan",
              "required": false,
              "schema": {
                "type": "integer",
                "example": 3
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Get History Transaksi berhasil",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 0
                      },
                      "message": {
                        "type": "string",
                        "example": "Get History Berhasil"
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "offset": {
                            "type": "integer",
                            "example": 0
                          },
                          "limit": {
                            "type": "integer",
                            "example": 3
                          },
                          "records": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "invoice_number": {
                                  "type": "string",
                                  "example": "INV17082023-001"
                                },
                                "transaction_type": {
                                  "type": "string",
                                  "example": "TOPUP"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "Top Up balance"
                                },
                                "total_amount": {
                                  "type": "integer",
                                  "example": 100000
                                },
                                "created_on": {
                                  "type": "string",
                                  "format": "date-time",
                                  "example": "2023-08-17T10:10:10.000Z"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 108
                      },
                      "message": {
                        "type": "string",
                        "example": "Token tidak valid atau kadaluwarsa"
                      },
                      "data": {
                        "type": "null"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }